organised the components and services in their respective folders
