+++ .hintrc
@@ -8,6 +8,12 @@
       {
         "label": "off"
       }
+    ],
+    "axe/name-role-value": [
+      "default",
+      {
+        "button-name": "off"
+      }
     ]
   }
 }
