+++ src/app/app.routes.ts
@@ -3,10 +3,10 @@ import { LoginComponent } from './login/login.component';
 import { ProfileComponent } from './profile/profile.component';
 import { RegisterComponent } from './register/register.component';
 import { AppComponent } from './app.component';
-import { CatalogueComponent } from './catalogue/catalogue.component';
+import { CatalogueComponent } from './Products/catalogue/catalogue.component';
 import { ResetComponent } from './reset/reset.component';
 import { UserDetailComponent } from './user-detail/user-detail.component';
-import { CartComponent } from './cart/cart.component';
+import { CartComponent } from './Carts/cart/cart.component';
 
 export const routes: Routes = [
   //{ path: '', component: AppComponent },
